+++ package.json
@@ -11,6 +11,7 @@
     "start": "npm run dev"
   },
   "dependencies": {
+    "axios": "^1.7.3",
     "fabric": "^5.3.0",
     "gapi-script": "^1.2.0",
     "jspdf": "^2.5.1",
@@ -20,6 +21,7 @@
     "react-icons": "^5.2.1",
     "react-joyride": "^2.8.1",
     "react-modal": "^3.16.1",
+    "react-router-dom": "^6.26.0",
     "react-simple-chatbot": "^0.5.0",
     "rollup": "^4.18.0",
     "svgcanvas": "^2.5.0"
